hal-win-task: fix integer conversion warning on msvc
add explicit cast to DWORD when passing usec to Sleep(). Sleep() takes
32-bit DWORD but usec is mrb_int which can be 64-bit, causing warning
c4244.

+++ mrbgems/hal-win-task/src/task_hal.c
@@ -129,7 +129,7 @@ mrb_hal_task_sleep_us(mrb_state *mrb, mrb_int usec)
 
   /* Windows Sleep() takes milliseconds, convert from microseconds */
   if (usec >= 0) {
-    Sleep(usec / 1000);
+    Sleep((DWORD)(usec / 1000));
   }
 }
